+++ tests/extension/executors/utils.test.ts
@@ -8,13 +8,11 @@ import { expect, vi, test, suite, beforeEach } from 'vitest'
 import { ENV_STORE } from '../../../src/extension/constants'
 import {
   retrieveShellCommand,
-  parseCommandSeq,
   getCellShellPath,
   getSystemShellPath,
   getCellCwd,
   isShellLanguage,
   getCmdShellSeq,
-  prepareCmdSeq,
   getCmdSeq,
 } from '../../../src/extension/executors/utils'
 import { getWorkspaceFolder, getAnnotations } from '../../../src/extension/utils'
@@ -222,132 +220,6 @@ test('supports multiline exports', async () => {
   expect(ENV_STORE.get('foo')).toBe('some\nmultiline\nexport')
   expect(ENV_STORE.get('foobar')).toBe('barfoo')
   expect(window.showInputBox).toBeCalledTimes(2)
-})
-
-suite('parseCommandSeq', () => {
-  beforeEach(() => {
-    vi.mocked(window.showInputBox).mockReset()
-  })
-
-  test('single-line export with prompt', async () => {
-    vi.mocked(window.showInputBox).mockImplementationOnce(async () => 'test value')
-
-    const res = await parseCommandSeq(['$ export TEST="<placeholder>"'].join('\n'), 'sh')
-
-    expect(res).toBeTruthy()
-    expect(res).toHaveLength(3)
-    expect(res?.[1]).toBe('export TEST="test value"')
-  })
-
-  test('single-line export with prompt disabled', async () => {
-    vi.mocked(window.showInputBox).mockImplementationOnce(async () => 'test value')
-
-    const res = await parseCommandSeq(['export TEST="placeholder"'].join('\n'), 'sh', false)
-
-    expect(window.showInputBox).toBeCalledTimes(0)
-
-    expect(res).toBeTruthy()
-    expect(res).toHaveLength(1)
-    expect(res![0]).toBe('export TEST="placeholder"')
-  })
-
-  test('single line export with cancelled prompt', async () => {
-    vi.mocked(window.showInputBox).mockImplementationOnce(async () => undefined)
-
-    const res = await parseCommandSeq(['export TEST="<placeholder>"'].join('\n'), 'sh')
-
-    expect(res).toBe(undefined)
-  })
-
-  test('non shell code with exports retains leading dollar signs', async () => {
-    vi.mocked(window.showInputBox).mockImplementationOnce(async () => undefined)
-
-    const res = await parseCommandSeq(
-      [
-        "$currentDateTime = date('Y-m-d H:i:s');",
-        '$fullGreeting = $greeting . " It\'s now " . $currentDateTime;',
-        'echo $fullGreeting;',
-      ].join('\n'),
-      'php',
-    )
-
-    expect(res).toBeTruthy()
-    expect(res).toStrictEqual([
-      "$currentDateTime = date('Y-m-d H:i:s');",
-      '$fullGreeting = $greeting . " It\'s now " . $currentDateTime;',
-      'echo $fullGreeting;',
-    ])
-  })
-
-  test('multiline export', async () => {
-    const exportLines = ['export TEST="I', 'am', 'doing', 'well!"']
-
-    const res = await parseCommandSeq(exportLines.join('\n'), 'sh')
-
-    expect(res).toBeTruthy
-    expect(res).toHaveLength(4)
-    expect(res).toStrictEqual(exportLines)
-  })
-
-  test('exports between normal command sequences', async () => {
-    vi.mocked(window.showInputBox).mockImplementationOnce(async () => 'test value')
-
-    const cmdLines = [
-      'echo "Hello!"',
-      'echo "Hi!"',
-      'export TEST="<placeholder>"',
-      'echo $TEST',
-      'export TEST_MULTILINE="This',
-      'is',
-      'a',
-      'multiline',
-      'env!"',
-      'echo $TEST_MULTILINE',
-    ]
-
-    const res = await parseCommandSeq(cmdLines.join('\n'), 'sh')
-
-    expect(res).toBeTruthy()
-    expect(res).toStrictEqual([
-      'echo "Hello!"',
-      'echo "Hi!"',
-      'export TEST="test value"',
-      '',
-      'echo $TEST',
-      ...['export TEST_MULTILINE="This', 'is', 'a', 'multiline', 'env!"'],
-      'echo $TEST_MULTILINE',
-    ])
-  })
-
-  // todo(seb): Not being used directly anywhere - deactivate for now and remove in future
-  // test('exports between normal command sequences with getCmdSeq', async () => {
-  //   vi.mocked(window.showInputBox).mockImplementationOnce(async () => 'test value')
-
-  //   const cmdLines = [
-  //     'echo "Hello!"',
-  //     'echo "Hi!"',
-  //     'export TEST="<placeholder>"',
-  //     'echo $TEST',
-  //     'export TEST_MULTILINE="This',
-  //     'is',
-  //     'a',
-  //     'multiline',
-  //     'env!"',
-  //     'echo $TEST_MULTILINE',
-  //   ]
-
-  //   const res = await parseCommandSeq(cmdLines.join('\n'), true, undefined, getCmdSeq) // signature obsolete
-
-  //   expect(res).toBeTruthy()
-  //   expect(res).toStrictEqual([
-  //     'echo "Hello!"',
-  //     'echo "Hi!"',
-  //     'export TEST="test value"',
-  //     'echo $TEST',
-  //     ...['export TEST_MULTILINE="This', 'is', 'a', 'multiline', 'env!"'],
-  //     'echo $TEST_MULTILINE',
-  //   ])
-  // })
 })
 
 suite('getCellShellPath', () => {
@@ -602,14 +474,6 @@ suite('getCmdShellSeq', () => {
   test('leading prompts', () => {
     const cellText = '$ docker build -t runme/demo .\n$ docker ps -qa'
     expect(getCmdShellSeq(cellText, 'darwin')).toMatchSnapshot()
-  })
-})
-
-suite('prepareCmdSeq', () => {
-  test('should eliminate leading dollar signs', () => {
-    expect(prepareCmdSeq('$ echo hi')).toStrictEqual(['echo hi'])
-    expect(prepareCmdSeq('  $  echo hi')).toStrictEqual(['echo hi'])
-    expect(prepareCmdSeq('echo 1\necho 2\n $ echo 4')).toStrictEqual(['echo 1', 'echo 2', 'echo 4'])
   })
 })
